pyproject.toml
[tool.poetry]
name = "haar_cascade_classifier"
version = "0.1.0"
description = ""
authors = ["Rafael Santos <rafaelsantos32141@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"

[tool.poetry.scripts]
preprocess = "haar_cascade_classifier.preprocessing:run"
train = "haar_cascade_classifier.train:run"
generate_xml = "haar_cascade_classifier.generate_xml:run"
test = "haar_cascade_classifier.test_classifier:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
